server/requirements.txt
capstone==4.0.2
diff_match_patch==20230430
Django==4.2.4
djangorestframework==3.14.0
gcloud==0.17.0
intervaltree==3.1.0
loguru==0.7.1
oauth2client==3.0.0
pyelftools==0.29
python-dotenv==1.0.0
PyYAML==6.0
Requests==2.31.0
django-cors-headers==4.2.0
httplib2==0.15.0
google-api-python-client==1.6